Fix prod version bug and change material-ui imports
Material-ui docs suggest that imports should be handled this way for
speed reasons. Also fixing bug where handling styles in different
module caused base theme not showing.
