+++ src/components/nav.js
@@ -1,8 +1,33 @@
 //@flow
 import React from "react"
-import useStyles from "styles"
-import { AppBar, Toolbar, Button, Link, IconButton } from "@material-ui/core"
+import AppBar from "@material-ui/core/AppBar"
+import Toolbar from "@material-ui/core/Toolbar"
+import Button from "@material-ui/core/Button"
+import Link from "@material-ui/core/Link"
+import IconButton from "@material-ui/core/IconButton"
 import HomeIcon from "@material-ui/icons/Home"
+import { makeStyles } from "@material-ui/core/styles"
+
+const useStyles = makeStyles(theme => ({
+  appBar: {
+    position: "relative",
+    borderBottom: `1px solid ${theme.palette.divider}`,
+    alignItems: "flex-end",
+    color: "rgb(72,72,72)",
+    backgroundColor: "white",
+  },
+  link: {
+    margin: theme.spacing(1, 1.5),
+    color: "inherit",
+  },
+  linkButton: {
+    margin: theme.spacing(1, 1.5),
+    color: "white",
+    padding: "10px 20px",
+    backgroundColor: "rgb(121, 131, 204)",
+    borderRadius: "20px",
+  },
+}))
 
 const Nav = () => {
   const classes = useStyles()
